Added the missing code bit
Necessary to add the full code snippet, as from the last time that code was used 1 variable has to be updated and without adding that you get compile errors. The variable to change isn't specified in the blog. SO the best way to overcome that is to just give the user full code snippet.